backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==10.1.0
environs[django]==9.3.2
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.3
requests==2.31.0
geopy==2.4.1
rollbar==1.0.0
psycopg2-binary==2.9.9
phonenumbers==8.13.24
gunicorn==21.2.0